Move remaining Node components
