+++ core/ui/build.gradle.kts
@@ -24,6 +24,7 @@ android { namespace = "org.meshtastic.core.ui" }
 
 dependencies {
     implementation(projects.core.database)
+    implementation(projects.core.model)
     implementation(projects.core.proto)
     implementation(projects.core.strings)
     implementation(libs.bundles.markdown)
